Move collapsed-item under _common
